requirements.txt
praw==7.6.0
pymongo==4.5.0
beautifulsoup4==4.12.2
selenium==4.14.0
pytesseract==0.3.10
Pillow==10.0.0
requests==2.31.0
scikit-learn==1.3.0
nltk==3.8.1
numpy==1.24.3
